# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Este es un proyecto de prueba técnica. El challenge consiste en realizar un API REST con sus respectivos endpoints usando Go.",
        "title": "MOVEO.AI CHALLENGE",
        "contact": {
            "name": "Matias Leonel Lodigiani",
            "email": "lodigiani.matias97@gmail.com"
        }
    },
    "host": "localhost:8080",
    "paths": {
        "/tasks": {
            "get": {
                "description": "Obtiene la lista completa de tareas",
                "tags": [
                    "tasks"
                ],
                "summary": "Obtener",
                "responses": {}
            },
            "post": {
                "description": "Añade una nueva tarea a la lista",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Crear",
                "parameters": [
                    {
                        "description": "Task a crear",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Obtiene la tarea por id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Obtener",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "description": "Modifica la tarea por id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Actualizar",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task a actualizar",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Task"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina la tarea por id",
                "tags": [
                    "tasks"
                ],
                "summary": "Eliminar",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.Task": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}